refactor: move wkt function to dedicated module
